Remove embedded web extension in install manifest.
This causes the embedded web extension to be parsed at startup, and
triggers a race condition in existing Firefox code during startup
on a clean profile between AddonManager and devtools code.

Removing this is not an issue for Screenshots because it delays startup
of the embedded web extension until "sessionstore-windows-restored" is
observed anyway.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1356394 for more info.