Stop updating tab state on tab activated
According to kmag, webextension buttons have per-tab state, so setting
the button enabled or disabled on update should be enough. Updating the
icon on activated may be causing performance issues.

This change does introduce a bug with the special first-run icon state:

1. have a few tabs open
2. enable screenshots
3. all tabs with screenshot-enabled URLs will show the first-run button
4. click the screenshot button in a given tab
5. close onboarding, notice that the tab's icon updates to the regular
enabled state
6. switch to other enabled tabs, and the icon will still be the special
first-run icon, not the regular enabled icon

We can probably work around this in the onboarding code, if we land this
patch in master.