Move the context menu item to bootstrap
Need to put the context menu in the right spot, at the very bottom of
the context menu. Probably need to create a menuseparator and a
menuitem.

Also need to call the initialization code via the 'onCommand'
attribute of the context menu item. XUL is funny. Currently just alerts
into the page when it's clicked.