Use the Photon page action when supported, the WebExtension browser action when not.
Extend the Photon-related port used between bootstrap.js and the
WebExtension so that bootstrap.js can tell the WebExtension
whether it's OK to use the Photon page action.  The WebExtension
should not attempt to use the browser action when Photon is
enabled because bootstrap.js will have removed the browser
action's navbar button.

Modify the test so that it checks the Photon page action when
Photon is enabled and the browser action when it's not.