Replace the WebExtension browser action with a Photon page action. (#3239)
* Replace the WebExtension browser action with a Photon page action.

This removes the browser action and adds a Photon page action,
pursuant to https://bugzilla.mozilla.org/show_bug.cgi?id=1366041.
Right now Photon page actions are unrelated to WebExtension page
actions unfortunately, which means that this patch has to do most
of its work in bootstrap.js.  The WebExtension part passes
messages to bootstrap.js to handle clicks and update the action's
title and icon as necessary.

* Test fix: Replace the WebExtension browser action with a Photon page action.

Fix the test for the previous commit.

* Update SHOOTER_BUTTON_ID in test.js.

* Use the Photon page action when supported, the WebExtension browser action when not.

Extend the Photon-related port used between bootstrap.js and the
WebExtension so that bootstrap.js can tell the WebExtension
whether it's OK to use the Photon page action.  The WebExtension
should not attempt to use the browser action when Photon is
enabled because bootstrap.js will have removed the browser
action's navbar button.

Modify the test so that it checks the Photon page action when
Photon is enabled and the browser action when it's not.

* Fix the expected button label in test.js for the Photon page action.

* Maybe fix PageActions eslint errors.