Raven cleanups (#3594)
* When sending pageview events for direct image views, only send the referrer origin

This matches similar logic in ga-activation.js that also scrubs the referrer
down to only the origin.

* Fix #3483, remove shot URLs from Sentry reports

This scrubs both shot URLs in referrers, as well as shot URLs from errors on the shot pages themselves. The paths are translated to /a-shot/redacted in this case.

This also adds a req argument to captureRavenException, which adds request-specific information to error reports (something that sendRavenMessage already did).