Fix #3581, make Screenshots work with third party cookies disabled
This adds a second attempt to login to wantsauth logins, one that runs in sitehelper.js, and tries to get the cookie set on a request that appears to come from the content page itself.

Note this does not firmly protect from the content page overwriting window.XMLHttpRequest and having the add-on use that object.