WIP for #3600, warn about failed login due to third party cookies
This adds backupCookieRequest to the sitehelper login process, to tell the site if third party cookies SHOULD work. If the site sees that third party cookies might not be enabled, then it does a second check to GET /api/set-login-cookie?check=1. If that request shows the cookie isn't set, then it changes the model to warn the user.

TODO: set something on My Shots too
TODO: write a proper LoginFailedWarning language/design
TODO: remove the changes that disable third party cookie support (it's disabled to make this easier to test)