Cancel some user timing measurements when page visibility changes to 'hidden'
JS background budgeting in Firefox 58 might cause Screenshots execution
to slow down if the user switches tabs while waiting for the UI to
update.

Exclude cases where the user timing event ends with an upload, since the
result of a successful upload is to switch tabs (to the shot page of the
uploaded shot).

Fixes #4055.