Migrate Screenshots to pure WebExtension
* The addons system, not the Screenshots addon, will be responsible for
detecting changes to the 'extensions.screenshots.disabled' pref, and
starting up or shutting down the Screenshots webextension when changes
occur (bug 1488971).

* The LibraryButton code has been moved from bootstrap.js into the
embedded API. We'll be able to remove this code as soon as Library
menu items can be declared from the WebExtension manifest (bug 1366389).

* Move addon/webextension to webextension, update all the refs.

* Update makefile and build tools.

* Remove the script that deletes duplicate locales (en_CA) at build
time, because this breaks running with web-ext, and there is code to
handle duplicate locales inside export_mc.py already.

* Keep `jpm` because `addons-linter` seems to depend on it

* Still TODO: README files and docs directory not yet updated