Upgrade to fluent 0.8.0, fluent-react 0.8.1 (#4935)
* Upgrade to fluent 0.8.0, fluent-react 0.8.1

* Add the attrs prop to Localized where needed

Starting with fluent-react 0.6.0, <Localized> requires the attrs prop to
be passed in order to set any localized attributes as props on the
wrapped component. attrs should be an object with attribute names as
keys and booleans as values. For instance:

    <Localized id="type-name" attrs={{placeholder: true}}>
        <input
            type="text"
            placeholder="Localizable placeholder"
            value={name}
            onChange={…}
        />
    </Localized>

* Use a server-side markup parser

* Use Localized overlay for TimeDiffs

<Localized> can take other React elements as props. If an element of the
same name is then found in the translation, the element passed as prop
is inserted into the translation (keeping its props) and its children
are set to the text content of the element found in the translation.

* Use Localized overlay in AdBanner